// src/app/components/code-viewer/interfaces/languages.enum.ts
export const enum Language {
  Bash = 'bash',
  C = 'c',
  CSharp = 'csharp',
  CPluPlus = 'cpluplus',
  CSS = 'css',
  Diff = 'diff',
  Basic = 'basic',
  DotNET = 'dotnet',
  Go = 'go',
  GraphQL = 'graphql',
  HTML = 'html',
  XML = 'xml',
  JSON = 'json',
  Java = 'java',
  JavaScript = 'javascript',
  Kotlin = 'kotlin',
  Less = 'less',
  Lua = 'lua',
  Makefile = 'makefile',
  Markdown = 'markdown',
  ObjectiveC = 'objectivec',
  PHP = 'php',
  Template = 'template',
  Perl = 'perl',
  Plain = 'plain',
  Python = 'python',
  REPL = 'repl',
  R = 'r',
  Ruby = 'ruby',
  Rust = 'rust',
  SCSS = 'scss',
  SQL = 'sql',
  Shell = 'shell',
  Session = 'session',
  Swift = 'swift',
  TOML = 'toml',
  INI = 'ini',
  TypeScript = 'typescript',
  Visual = 'visual',
  WebAssembly = 'webassembly',
  YAML = 'yaml',
}